requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.8.2
sqlalchemy==2.0.32
alembic==1.13.2
python-multipart==0.0.9
boto3==1.34.162
pandas==2.2.2
cachetools==5.3.3
python-dotenv==1.0.1